🎨style: add scroll up animation
