+++ app/components/scrollToTop.tsx
@@ -15,12 +15,12 @@ export function ScrollToTop(): React.JSX.Element {
   return (
     <button
       type="button"
-      className={`btn btn-square btn-primary fixed right-10 bottom-10 gap-0 ${
+      className={`btn btn-square btn-primary fixed right-10 bottom-10 gap-0 hover:scale-110 ${
         scrollY < scrollPoint ? "hidden" : "fade-in-up"
       }`}
       onClick={() => window.scrollTo({ top: 0, behavior: "smooth" })}
     >
-      <ChevronDoubleUpIcon className="size-8" />
+      <ChevronDoubleUpIcon className="size-8 scroll-up" />
       TOP
     </button>
   )
