+++ app/globals.css
@@ -6,6 +6,10 @@
   animation: fade-in-up 1.5s;
 }
 
+.scroll-up {
+  animation: scroll-up 1.5s infinite;
+}
+
 .tilt-shaking:hover {
   animation: tilt-shaking 0.25s linear infinite;
 }
@@ -18,6 +22,18 @@
   to {
     opacity: 1;
     transform: translatey(0);
+  }
+}
+
+@keyframes scroll-up {
+  0% {
+    transform: translate(0);
+  }
+  20% {
+    transform: translatey(-8px);
+  }
+  40% {
+    transform: translate(0);
   }
 }
