+++ app/lib/constant.ts
@@ -1,0 +1,65 @@
+import type { Navigation } from "@/app/interfaces/navigation"
+import type { SiteUrl } from "@/app/interfaces/siteUrl"
+import {
+  HeartIcon,
+  NewspaperIcon,
+  QuestionMarkCircleIcon,
+  SparklesIcon,
+  TrophyIcon,
+  TruckIcon,
+} from "@heroicons/react/24/outline"
+
+export const SITE_TITLE: string = "こどもテックキャラバン"
+
+const nextArea: SiteUrl[] = [
+  { name: "広島", href: "/hiroshima" },
+  { name: "千葉", href: "/chiba" },
+] as const
+const history: SiteUrl[] = [{ name: "広島", href: "/history" }] as const
+const helper: SiteUrl[] = [
+  { name: "サポーター", href: "/supporter" },
+  { name: "パートナー", href: "/partner" },
+] as const
+const siteInfo: SiteUrl[] = [
+  { name: "FAQ", href: "/faq" },
+  { name: "個人情報保護方針", href: "/privacy-policy" },
+] as const
+
+export const navItems: Navigation[] = [
+  {
+    name: "最新情報",
+    icon: NewspaperIcon,
+    color: "text-primary",
+    href: "/#news",
+  },
+  {
+    name: "活動紹介",
+    icon: SparklesIcon,
+    color: "text-warning",
+    href: "/#about",
+  },
+  {
+    name: "開催予定",
+    icon: TruckIcon,
+    color: "text-accent",
+    content: [...nextArea],
+  },
+  {
+    name: "過去の実績",
+    icon: TrophyIcon,
+    color: "text-error",
+    content: [...history],
+  },
+  {
+    name: "サポーター/パートナー",
+    icon: HeartIcon,
+    color: "text-secondary",
+    content: [...helper],
+  },
+  {
+    name: "お問い合わせ",
+    icon: QuestionMarkCircleIcon,
+    color: "text-info",
+    content: [...siteInfo],
+  },
+] as const
